+++ fiddles/d3/fiddle-0026-Nvd3VerticalBarModel/src/meta.js
@@ -1,0 +1,7 @@
+let metadata = {
+    fiddleHeader: 'Nvd3 - Vertical Bar Model',
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/d3/fiddle-0026-Nvd3VerticalBarModel'
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};