+++ CHANGELOG.markdown
@@ -812,9 +812,16 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **D3**
     1.  Added [fiddles/d3/fiddle-0023-ElementTransition](fiddles/d3/fiddle-0023-ElementTransition)
     2.  Added [fiddles/d3/fiddle-0024-BarChart](fiddles/d3/fiddle-0024-BarChart)
+    3.  Started [fiddles/d3/fiddle-0025-DiscreteBarChart](fiddles/d3/fiddle-0025-DiscreteBarChart)
+    4.  Started [fiddles/d3/fiddle-0026-Nvd3VerticalBarModel](fiddles/d3/fiddle-0026-Nvd3VerticalBarModel)
+    5.  Added [fiddles/d3/fiddle-0027-Level3Events](fiddles/d3/fiddle-0027-Level3Events)
 * **Angular2**
     1.  Started [fiddles/angular2/fiddle-0031-PlayerToolbar](fiddles/angular2/fiddle-0031-PlayerToolbar)
 * **Rx**
     1.  Refactored [fiddles/rxjs/fiddle-0001-AjaxObservable](fiddles/rxjs/fiddle-0001-AjaxObservable)
     2.  Added [fiddles/rxjs/fiddle-0002-TimeObservable](fiddles/rxjs/fiddle-0002-TimeObservable)
-    3.  Added [fiddles/rxjs/fiddle-0003-Disposable](fiddles/rxjs/fiddle-0003-Disposable)
+    3.  Started [fiddles/rxjs/fiddle-0003-Disposable](fiddles/rxjs/fiddle-0003-Disposable)
+* **Node**
+    1.  Started [fiddles/node/fiddle-0002-FileSystemApi](fiddles/node/fiddle-0002-FileSystemApi)
+* **Bash**
+    1.  Added [fiddles/bash/fiddle-0073-CsvToJson](fiddles/bash/fiddle-0073-CsvToJson)